refactor: 도메인 패키지로 분리
