+++ backend/src/main/java/zipgo/petfood/controller/PetFoodController.java
@@ -1,4 +1,4 @@
-package zipgo.controller;
+package zipgo.petfood.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import zipgo.controller.dto.GetPetFoodsResponse;
-import zipgo.domain.PetFood;
-import zipgo.service.PetFoodService;
+import zipgo.petfood.controller.dto.GetPetFoodsResponse;
+import zipgo.petfood.domain.PetFood;
+import zipgo.petfood.service.PetFoodService;
 
 @RestController
 @RequestMapping("/pet-foods")
