+++ backend/src/test/java/zipgo/domain/repository/KeywordRepositoryTest.java
@@ -6,7 +6,8 @@ import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import zipgo.domain.Keyword;
+import zipgo.petfood.domain.Keyword;
+import zipgo.petfood.domain.repository.KeywordRepository;
 
 @DataJpaTest
 class KeywordRepositoryTest {
